+++ CHANGELOG.md
@@ -120,3 +120,24 @@ Version 1.2.3 - improvements and some minor bug fixes.
 
 🛠️ Release Patch : 
 - [x] Now using brace `{}` as delimiter in settings is prohibited. 
+
+
+### Version 1.2.4  
+Version 1.2.4 - bug fixes and minor improvements. 
+
+🐛 Fixed Bugs :
+- [x] Fix [bug #34](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/34) now amrkdonw export can work correctly for escaped dollar sign. 
+
+- [x] **(Critical)** Fix [Critical #35](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/35), when no tag to rename and delete unused tags is enabled, the plugin will delete unused tags normally😅.  
+
+- [x] Fix [bug #31](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/31), now citations can rendered correctly both in outline preview and embedded link preview. 
+
+🚀 Enhancements : 
+- [x] add [enhancement #32](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/32), file superscripts will be rendered as footnotes in pdf export. Also change default citation color as same as file superscript color. 
+
+- [x] open the exported file after export pdf command.  
+
+- [x] set default equation widget width to 500px. 
+
+- [x] better empty citation rendering (now empty citations will be rendered as `()` instead of ``)
+